Add iconName util and helper to get the icon name based of a file extension
