+++ app/components/utils/documents/file-type-pill.hbs
@@ -1,7 +1,7 @@
 <Auk::Pill
   data-test-file-type-pill
   @skin="white"
-  @icon={{this.icon}}
+  @icon={{icon-name @extension}}
 >
   {{uppercase @extension}}
 </Auk::Pill>
